includes _pb2

+++ kgekit/data.py
@@ -14,8 +14,8 @@ def split_golden_set(triples, valid_ratio, test_ratio):
     """Split the data into train/valid/test sets."""
     assert valid_ratio >= 0.0
     assert test_ratio >= 0.0
-    num_valid = len(triples) * valid_ratio
-    num_test = len(triples) * test_ratio
+    num_valid = int(len(triples) * valid_ratio)
+    num_test = int(len(triples) * test_ratio)
     valid_set = triples[:num_valid]
     test_set = triples[num_valid:num_valid+num_test]
     train_set = triples[num_valid+num_test:]
